backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
pillow==12.2.0
sqlparse==0.5.5
tzdata==2026.1
gunicorn==23.0.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
